# fix node position is not correct