@holycanvas 2022-05-14
Impact on performance：

test case: 2D Sprite Translation, device: vivo

2000 sprite: 
before: 43 fps
after: 41-42 fps

4100 sprite:
before: 23 fps
after: 22 fps

6000 sprites:
before: 16 fps
after: 16 fps